+++ backend/src/config/dto/adminConfig.dto.ts
@@ -9,6 +9,9 @@ export class AdminConfigDTO extends ConfigDTO {
   secret: boolean;
 
   @Expose()
+  defaultValue: string;
+
+  @Expose()
   updatedAt: Date;
 
   @Expose()